+++ src/kreyolib/convert/__init__.py
@@ -1,0 +1,4 @@
+from kreyolib.convert.num_to_text import num_to_text
+from kreyolib.convert.text_to_num import text_to_num
+
+__all__ = ["num_to_text", "text_to_num"]